fix comment and change Id type to long

+++ src/ChatWootApi/Application/Models/ContactCreatePayload.cs
@@ -9,61 +9,61 @@ namespace ChatWootApi.Application.Models;
 public sealed record ContactCreatePayload
 {
     /// <summary>
-    /// 收件箱ID。
+    /// 联系人所属收件箱ID
     /// </summary>
     [JsonPropertyName("inbox_id")]
-    public decimal? InboxId { get; set; }
+    public long? InboxId { get; set; }
 
     /// <summary>
-    /// 名称。
+    /// 联系人姓名
     /// </summary>
     [JsonPropertyName("name")]
     public string? Name { get; set; }
 
     /// <summary>
-    /// 邮箱。
+    /// 联系人的电子邮件
     /// </summary>
     [JsonPropertyName("email")]
     public string? Email { get; set; }
 
     /// <summary>
-    /// Blocked。
+    /// 联系人是否被屏蔽
     /// </summary>
     [JsonPropertyName("blocked")]
     public bool? Blocked { get; set; }
 
     /// <summary>
-    /// 电话号码。
+    /// 联系人的电话号码
     /// </summary>
     [JsonPropertyName("phone_number")]
     public string? PhoneNumber { get; set; }
 
     /// <summary>
-    /// 头像。
+    /// 使用头像图像二进制文件发送表单数据或使用 avatar_url
     /// </summary>
     [JsonPropertyName("avatar")]
     public string? Avatar { get; set; }
 
     /// <summary>
-    /// 头像URL。
+    /// 联系人头像的 jpeg、png 文件的 url
     /// </summary>
     [JsonPropertyName("avatar_url")]
     public string? AvatarUrl { get; set; }
 
     /// <summary>
-    /// 标识符。
+    /// 联系人在外部系统中的唯一标识符
     /// </summary>
     [JsonPropertyName("identifier")]
     public string? Identifier { get; set; }
 
     /// <summary>
-    /// 附加属性。
+    /// 您可以在其中存储联系人的附加属性的对象。示例{“类型”：“客户”，“年龄”：30}
     /// </summary>
     [JsonPropertyName("additional_attributes")]
     public IDictionary<string, object>? AdditionalAttributes { get; set; }
 
     /// <summary>
-    /// 自定义属性。
+    /// 您可以在其中存储联系人的自定义属性的对象。例如 {"type":"customer", "age":30}，这应该具有有效的自定义属性定义。
     /// </summary>
     [JsonPropertyName("custom_attributes")]
     public IDictionary<string, object>? CustomAttributes { get; set; }
